feat(frontend): show disks and devices in machines/machine page
Show disks and devices for machines in the machines/machine page, even for maintenance mode machines.
