+++ frontend/e2e/qemu/clusters.spec.ts
@@ -4,7 +4,6 @@
 // included in the LICENSE file.
 import fs from 'node:fs/promises'
 import os from 'node:os'
-import path from 'node:path'
 
 import type { Page } from '@playwright/test'
 import { milliseconds } from 'date-fns'
@@ -23,11 +22,9 @@ const test = base.extend<{ saveSupportBundleOnFailure: void }>({
 
       // Save support bundle if the test failed
       if (testInfo.status !== 'passed') {
-        const bundleDir = path.resolve(testInfo.outputDir, '..', 'support-bundles')
-        const bundlePath = path.join(bundleDir, `support-bundle-${clusterName}.zip`)
+        const bundlePath = testInfo.outputPath(`support-bundle-${clusterName}.zip`)
 
         try {
-          await fs.mkdir(bundleDir, { recursive: true })
           await omnictl(['support', '--cluster', clusterName, '--output', bundlePath])
           await testInfo.attach(`support-bundle-${clusterName}.zip`, { path: bundlePath })
         } catch (e) {
@@ -42,6 +39,54 @@ const test = base.extend<{ saveSupportBundleOnFailure: void }>({
 test.describe.configure({ mode: 'serial', retries: 0 })
 
 const cpMachineName = 'deadbeef'
+
+test('machine overview tabs', async ({ page }) => {
+  await page.goto('/')
+  await page.getByRole('link', { name: 'Machines' }).click()
+
+  // Open first visible machine
+  await page
+    .getByRole('link', { name: /machine-[-a-f0-9]+/ })
+    .first()
+    .click()
+
+  await test.step('Validate logs tab', async () => {
+    await page.getByRole('tab', { name: 'Logs', exact: true }).click()
+
+    await page.getByPlaceholder('Search').fill('[talos] [initramfs] booting Talos')
+    await expect
+      .soft(page.getByText(/\[talos\] \[initramfs\] booting Talos v\d\.\d+\.\d+/).first())
+      .toBeVisible()
+
+    await page.getByPlaceholder('Search').clear()
+    await expect
+      .soft(page.getByText(/\[talos\] \[initramfs\] booting Talos v\d\.\d+\.\d+/))
+      .toBeHidden()
+  })
+
+  await test.step('Validate patches tab', async () => {
+    await page.getByRole('tab', { name: 'Patches', exact: true }).click()
+    await expect.soft(page.getByRole('heading', { name: 'No Config Patches' })).toBeVisible()
+  })
+
+  await test.step('Validate disks tab', async () => {
+    await page.getByRole('tab', { name: 'Disks', exact: true }).click()
+
+    await expect
+      .soft(page.getByRole('region', { name: 'vda' }).getByText('Unallocated6.00 GB'))
+      .toBeVisible()
+  })
+
+  await test.step('Validate devices tab', async () => {
+    await page.getByRole('tab', { name: 'Devices', exact: true }).click()
+    await expect.soft(page.getByText('Ethernet controller')).toBeVisible()
+  })
+
+  await test.step('Validate extensions tab', async () => {
+    await page.getByRole('tab', { name: 'Extensions', exact: true }).click()
+    await expect.soft(page.getByText('siderolabs/hello-world-service')).toBeVisible()
+  })
+})
 
 test('create cluster', async ({ page }) => {
   test.setTimeout(milliseconds({ minutes: 16 }))
@@ -310,7 +355,16 @@ test('node overview tabs', async ({ page }) => {
 
   await test.step('Validate console logs tab', async () => {
     await page.getByRole('tab', { name: 'Console Logs', exact: true }).click()
-    await expect(page.getByText('[talos]').first()).toBeVisible()
+
+    await page.getByPlaceholder('Search').fill('[talos] [initramfs] booting Talos')
+    await expect
+      .soft(page.getByText(/\[talos\] \[initramfs\] booting Talos v\d\.\d+\.\d+/).first())
+      .toBeVisible()
+
+    await page.getByPlaceholder('Search').clear()
+    await expect
+      .soft(page.getByText(/\[talos\] \[initramfs\] booting Talos v\d\.\d+\.\d+/))
+      .toBeHidden()
   })
 
   await test.step('Validate config tab', async () => {
