+++ frontend/src/pages/(authenticated)/machines/[machine]/extensions.vue
@@ -5,11 +5,11 @@ Use of this software is governed by the Business Source License
 included in the LICENSE file.
 -->
 <script setup lang="ts">
-import MachineDevices from '@/views/Machines/MachineDevices.vue'
+import MachineExtensions from '@/views/Machines/MachineExtensions.vue'
 
-definePage({ name: 'NodeDevices' })
+definePage({ name: 'MachineExtensions' })
 </script>
 
 <template>
-  <MachineDevices />
+  <MachineExtensions :machine-id="$route.params.machine" read-only />
 </template>
