+++ frontend/src/views/Machines/components/UpdateExtensionsModal.stories.ts
