+++ frontend/src/views/Machines/components/UpdateExtensionsModal.vue
