+++ frontend/src/views/Nodes/components/DiskUsageBar.vue
@@ -10,10 +10,7 @@ import { computed } from 'vue'
 import type { Resource } from '@/api/grpc'
 import { itemID } from '@/api/watch'
 import { formatBytes } from '@/methods'
-import type {
-  TalosDiscoveredVolumeSpec,
-  TalosDiskSpec,
-} from '@/pages/(authenticated)/clusters/[cluster]/machine/[machine]/disks.vue'
+import type { TalosDiscoveredVolumeSpec, TalosDiskSpec } from '@/views/Machines/MachineDisks.vue'
 
 const { disk, volumes } = defineProps<{
   disk: Resource<TalosDiskSpec>
